move Request factory functions into Request.h/cpp
