+++ src/MicroOcpp/Core/Request.h
@@ -116,6 +116,12 @@ public:
     bool isRequestSent();
 };
 
+/*
+ * Simple factory functions
+ */
+std::unique_ptr<Request> makeRequest(std::unique_ptr<Operation> op);
+std::unique_ptr<Request> makeRequest(Operation *op); //takes ownership of op
+
 } //end namespace MicroOcpp
 
  #endif
