+++ src/MicroOcpp/Core/RequestQueue.cpp
@@ -7,7 +7,6 @@
 #include <MicroOcpp/Core/RequestQueue.h>
 #include <MicroOcpp/Core/Request.h>
 #include <MicroOcpp/Core/Connection.h>
-#include <MicroOcpp/Core/SimpleRequestFactory.h>
 #include <MicroOcpp/Core/OcppError.h>
 #include <MicroOcpp/Core/OperationRegistry.h>
 #include <MicroOcpp/Operations/StatusNotification.h>
