+++ src/MicroOcpp/Model/Diagnostics/DiagnosticsService.cpp
@@ -5,7 +5,7 @@
 #include <MicroOcpp/Model/Diagnostics/DiagnosticsService.h>
 #include <MicroOcpp/Core/Context.h>
 #include <MicroOcpp/Model/Model.h>
-#include <MicroOcpp/Core/SimpleRequestFactory.h>
+#include <MicroOcpp/Core/Request.h>
 #include <MicroOcpp/Debug.h>
 
 #include <MicroOcpp/Operations/GetDiagnostics.h>
