+++ src/MicroOcpp/Model/Reset/ResetService.cpp
@@ -5,7 +5,7 @@
 #include <MicroOcpp/Model/Reset/ResetService.h>
 #include <MicroOcpp/Core/Context.h>
 #include <MicroOcpp/Model/Model.h>
-#include <MicroOcpp/Core/SimpleRequestFactory.h>
+#include <MicroOcpp/Core/Request.h>
 #include <MicroOcpp/Core/Configuration.h>
 #include <MicroOcpp/Operations/Reset.h>
