+++ tests/Certificates.cpp
@@ -13,7 +13,6 @@
 
 #include <MicroOcpp/Core/Context.h>
 #include <MicroOcpp/Operations/CustomOperation.h>
-#include <MicroOcpp/Core/SimpleRequestFactory.h>
 
 #include <MicroOcpp/Core/FilesystemAdapter.h>
 #include <MicroOcpp/Core/FilesystemUtils.h>
