+++ tests/ChargePointError.cpp
@@ -9,7 +9,6 @@
 
 #include <MicroOcpp/Core/Context.h>
 #include <MicroOcpp/Operations/CustomOperation.h>
-#include <MicroOcpp/Core/SimpleRequestFactory.h>
 #include <MicroOcpp/Debug.h>
 
 #include <MicroOcpp/Core/FilesystemAdapter.h>
